// apps/api/src/features/feedbacks/errors/feedback-error-messages.enum.ts
export enum FeedbackErrorMessages {
  FEEDBACK_NOT_FOUND = 'Feedback not found',
  FEEDBACK_ALREADY_EXISTS = 'You have already left feedback for this user',
  CANNOT_FEEDBACK_YOURSELF = 'You cannot leave feedback for yourself',
  CANNOT_MODIFY_OTHER_FEEDBACK = 'You cannot modify feedback from other users',
  RECEIVER_NOT_FOUND = 'Receiver user not found',
  CANNOT_VIEW_OTHER_FEEDBACKS = 'You cannot view feedbacks from other users',
}